Tweak lookahead/parse algorithm
Tweaked the lookahead and parse algorithm so that
datetimes such as 'tomorrow at 5pm' can be parsed
in a command.

Originally, the lookahead algorithm would
reverse the sentence temporarily and stop at
'at 5pm' because it believed it was a complete
input. It now starts at the front.

The `_parse_intype` algorithm would only look
at the beginning of the sentence and thus considered
'tomorrow' a complete input. This has bene changed
now so that it will look at the whole sentence and
if the whole sentence does not match in InputType,
it will pop out the first word and continue looking.